Move `extensions/pkg/log` to `pkg/logger`
